website/requirements.txt
celery==5.2.7
Django==4.1.1
nordigen==1.3.0
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
django-redis==5.2.0